move component files into sections for clarity
