+++ client/components/booking-details/PricePerNight.jsx
