+++ client/components/booking-details/Star.jsx
